1.修改用户重置密码 2.操作改为更多 3.操作提示统一改为弹出窗口提示

+++ src/view/system/role/index.vue
@@ -11,22 +11,15 @@
       <Table :data="list" :columns="tableColumns" :loading="tableLoading" border stripe>
         <template slot-scope="{ row, index }" slot="action">
             <Button  type="primary" size="small" style="margin: 5px" @click="addOrUpdateHandle(row.id)">编辑</Button>
-            <Poptip
-                confirm
-                transfer
-                title="您确定要删除吗?"
-                @on-ok="handleDelete(row.id)"
-                >
-               <Button type="error" size="small">删除</Button>
-            </Poptip>
             <Dropdown @on-click="dropDownClick($event,row)" transfer>
-              <Button type="primary" size="small" style="margin: 10px">
+              <Button type="warning" size="small" style="margin: 5px">
                 更多
                 <Icon type="ios-arrow-down"></Icon>
               </Button>
               <DropdownMenu slot="list">
                 <DropdownItem name="setRole">角色权限</DropdownItem>
                 <DropdownItem name="setDataRole">数据权限</DropdownItem>
+                <DropdownItem name="delete">删除</DropdownItem>
               </DropdownMenu>
             </Dropdown>
         </template>
@@ -140,11 +133,17 @@ export default {
       })
     },
     handleDelete(id) {
-      deleteRole(id).then(res => {
-        this.$Message.success(res.msg)
-        this.getList()
-      }).catch(error => {
-        this.$Message.error(error.msg)
+      this.$Modal.confirm({
+        title: '提示',
+        content: '确认要删除该数据吗？',
+        onOk: () => {
+          deleteRole(id).then(res => {
+            this.$Message.success(res.msg)
+            this.getList()
+          }).catch(error => {
+            this.$Message.error(error.msg)
+          })
+        }
       })
     },
     // 清空查询值的时候 重新加载列表数据
@@ -154,14 +153,15 @@ export default {
       }
     },
     dropDownClick(e, row) {
-      console.log(row)
-      console.log(e)
       switch (e) {
         case 'setRole':
           this.setRole(row)
           break
         case 'setDataRole':
           alert('setDataRole')
+          break
+        case 'delete':
+          this.handleDelete(row.id)
           break
       }
     }
